requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.5
flake8==7.0.0
flake8-html==0.4.3
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
